Update and rename average to introduce
I always try to write straightforward code, which makes it easier for other programmers to look at my code. I think the time and effort spent reading and understanding the code is often more than the time to sit optimally coding. So I guess I need to optimize the code to make it a module with DI and 100% coverage test.